// typings.d.ts
export type TweetBody = {
    text: string,
    username: string,
    profileImg: string,
    image?: string,
}

export interface ITweet extends TweetBody {
    _id: string,
    _createdAt: string,
    _updatedAt: string,
    _rev: string,
    _type: 'tweet',
    blockTweet: boolean
}

export type CommentBody = {
    comment: string;
    tweetId: string;
    username: string;
    profileImg: string;
}

export interface IComment extends CommentBody {
    _id: string,
    _createdAt: string,
    _updatedAt: string,
    _rev: string,
    _type: 'comment',
    tweet: {
        _ref: string;
        _type: 'reference'
    }
}